requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
requests>=2.31.0
langchain==0.1.0
langchain-community==0.0.10
openai-whisper==20231117
torch==2.1.0
sqlite-utils==3.35.2
streamlit==1.29.0
pydantic==2.5.0
python-multipart==0.0.6
audio-recorder-streamlit==0.0.8
pandas==2.1.3
